refactor(appsec): use neutral HTTP helper imports
Keep blocked-response behavior unchanged while avoiding direct AppSec dependencies on tracing product modules.
