+++ ddtrace/appsec/_contrib/flask/__init__.py
@@ -8,11 +8,6 @@ from typing import MutableMapping
 from typing import Optional
 from typing import cast
 
-from ddtrace._trace.http_semantics import set_method_tag
-from ddtrace._trace.http_semantics import set_query_string_tag
-from ddtrace._trace.http_semantics import set_status_code_tag
-from ddtrace._trace.http_semantics import set_url_tags_server
-from ddtrace._trace.http_semantics import user_agent_tag
 from ddtrace.appsec._asm_request_context import _call_waf_first
 from ddtrace.appsec._asm_request_context import _on_context_ended
 from ddtrace.appsec._asm_request_context import _set_headers_and_response
@@ -27,6 +22,11 @@ from ddtrace.appsec._asm_request_context import set_waf_address
 from ddtrace.appsec._utils import Block_config
 from ddtrace.contrib import trace_utils
 from ddtrace.contrib.internal.trace_utils_base import _get_request_header_user_agent
+from ddtrace.contrib.internal.trace_utils_base import set_method_tag
+from ddtrace.contrib.internal.trace_utils_base import set_query_string_tag
+from ddtrace.contrib.internal.trace_utils_base import set_status_code_tag
+from ddtrace.contrib.internal.trace_utils_base import set_url_tags_server
+from ddtrace.contrib.internal.trace_utils_base import user_agent_tag
 from ddtrace.internal import core
 from ddtrace.internal.appsec.prototypes import SpanProtocol
 from ddtrace.internal.constants import REQUEST_PATH_PARAMS
